+++ _wasm_crypto/src/digest.rs
@@ -1,23 +1,25 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 use digest::{
-  BlockInput, Digest, DynDigest, ExtendableOutput, Reset, Update,
-  VariableOutput,
+  core_api::BlockSizeUser, Digest, DynDigest, ExtendableOutput,
+  ExtendableOutputReset, Reset, Update,
 };
+use typenum::{U32, U48};
 
 /// Enum wrapper over all supported digest implementations.
 #[derive(Clone)]
 pub enum Context {
-  Blake2b(Box<blake2::Blake2b>),
-  Blake2b256(Box<blake2::VarBlake2b>),
-  Blake2b384(Box<blake2::VarBlake2b>),
-  Blake2s(Box<blake2::Blake2s>),
+  Blake2b(Box<blake2::Blake2b512>),
+  Blake2b256(Box<blake2::Blake2b<U32>>),
+  Blake2b384(Box<blake2::Blake2b<U48>>),
+  Blake2s(Box<blake2::Blake2s256>),
   Blake3(Box<blake3::Hasher>),
   Keccak224(Box<sha3::Keccak224>),
   Keccak256(Box<sha3::Keccak256>),
   Keccak384(Box<sha3::Keccak384>),
   Keccak512(Box<sha3::Keccak512>),
+  Md4(Box<md4::Md4>),
   Md5(Box<md5::Md5>),
-  Ripemd160(Box<ripemd160::Ripemd160>),
+  Ripemd160(Box<ripemd::Ripemd160>),
   Sha1(Box<sha1::Sha1>),
   Sha3_224(Box<sha3::Sha3_224>),
   Sha3_256(Box<sha3::Sha3_256>),
@@ -29,6 +31,7 @@ pub enum Context {
   Sha512(Box<sha2::Sha512>),
   Shake128(Box<sha3::Shake128>),
   Shake256(Box<sha3::Shake256>),
+  Tiger(Box<tiger::Tiger>),
 }
 
 use Context::*;
@@ -37,18 +40,15 @@ impl Context {
   pub fn new(algorithm_name: &str) -> Result<Context, &'static str> {
     Ok(match algorithm_name {
       "BLAKE2B" => Blake2b(Default::default()),
-      "BLAKE2B-256" => {
-        Blake2b256(Box::new(blake2::VarBlake2b::new(256 / 8).unwrap()))
-      }
-      "BLAKE2B-384" => {
-        Blake2b384(Box::new(blake2::VarBlake2b::new(384 / 8).unwrap()))
-      }
+      "BLAKE2B-256" => Blake2b256(Default::default()),
+      "BLAKE2B-384" => Blake2b384(Default::default()),
       "BLAKE2S" => Blake2s(Default::default()),
       "BLAKE3" => Blake3(Default::default()),
       "KECCAK-224" => Keccak224(Default::default()),
       "KECCAK-256" => Keccak256(Default::default()),
       "KECCAK-384" => Keccak384(Default::default()),
       "KECCAK-512" => Keccak512(Default::default()),
+      "MD4" => Md4(Default::default()),
       "MD5" => Md5(Default::default()),
       "RIPEMD-160" => Ripemd160(Default::default()),
       "SHA-1" => Sha1(Default::default()),
@@ -62,6 +62,7 @@ impl Context {
       "SHA-512" => Sha512(Default::default()),
       "SHAKE128" => Shake128(Default::default()),
       "SHAKE256" => Shake256(Default::default()),
+      "TIGER" => Tiger(Default::default()),
       _ => return Err("unsupported algorithm"),
     })
   }
@@ -70,7 +71,7 @@ impl Context {
   pub fn input_block_length(&self) -> usize {
     // For algorithm types that implement BlockInput and have a statically
     // available BlockSize as part of their type definition, we use that value.
-    fn static_block_length<T: BlockInput>(_: &T) -> usize {
+    fn static_block_length<T: BlockSizeUser>(_: &T) -> usize {
       <T::BlockSize as typenum::Unsigned>::to_usize()
     }
 
@@ -79,11 +80,12 @@ impl Context {
       Blake2b256(context) => static_block_length(&**context),
       Blake2b384(context) => static_block_length(&**context),
       Blake2s(context) => static_block_length(&**context),
-      Blake3(context) => static_block_length(&**context),
+      Blake3(_) => 64,
       Keccak224(context) => static_block_length(&**context),
       Keccak256(context) => static_block_length(&**context),
       Keccak384(context) => static_block_length(&**context),
       Keccak512(context) => static_block_length(&**context),
+      Md4(context) => static_block_length(&**context),
       Md5(context) => static_block_length(&**context),
       Ripemd160(context) => static_block_length(&**context),
       Sha1(context) => static_block_length(&**context),
@@ -95,6 +97,7 @@ impl Context {
       Sha256(context) => static_block_length(&**context),
       Sha384(context) => static_block_length(&**context),
       Sha512(context) => static_block_length(&**context),
+      Tiger(context) => static_block_length(&**context),
 
       // https://doi.org/10.6028/NIST.FIPS.202 specifies that:
       // - In general, the input block size (in bits) of a sponge function is
@@ -123,6 +126,7 @@ impl Context {
       Keccak256(context) => context.output_size(),
       Keccak384(context) => context.output_size(),
       Keccak512(context) => context.output_size(),
+      Md4(context) => context.output_size(),
       Md5(context) => context.output_size(),
       Ripemd160(context) => context.output_size(),
       Sha1(context) => context.output_size(),
@@ -134,6 +138,7 @@ impl Context {
       Sha256(context) => context.output_size(),
       Sha384(context) => context.output_size(),
       Sha512(context) => context.output_size(),
+      Tiger(context) => context.output_size(),
 
       // https://doi.org/10.6028/NIST.FIPS.202's Table 4 indicates that in order
       // to reach the target security strength for these algorithms, the output
@@ -167,6 +172,7 @@ impl Context {
       Keccak256(_) => "KECCAK-256",
       Keccak384(_) => "KECCAK-384",
       Keccak512(_) => "KECCAK-512",
+      Md4(_) => "MD5",
       Md5(_) => "MD5",
       Ripemd160(_) => "RIPEMD-160",
       Sha1(_) => "SHA-1",
@@ -180,6 +186,7 @@ impl Context {
       Sha512(_) => "SHA-512",
       Shake128(_) => "SHAKE128",
       Shake256(_) => "SHAKE256",
+      Tiger(_) => "TIGER",
     }
   }
 
@@ -194,6 +201,7 @@ impl Context {
       Keccak256(context) => Reset::reset(&mut **context),
       Keccak384(context) => Reset::reset(&mut **context),
       Keccak512(context) => Reset::reset(&mut **context),
+      Md4(context) => Reset::reset(&mut **context),
       Md5(context) => Reset::reset(&mut **context),
       Ripemd160(context) => Reset::reset(&mut **context),
       Sha1(context) => Reset::reset(&mut **context),
@@ -207,20 +215,22 @@ impl Context {
       Sha512(context) => Reset::reset(&mut **context),
       Shake128(context) => Reset::reset(&mut **context),
       Shake256(context) => Reset::reset(&mut **context),
+      Tiger(context) => Reset::reset(&mut **context),
     };
   }
 
   pub fn update(&mut self, data: &[u8]) {
     match self {
       Blake2b(context) => Digest::update(&mut **context, data),
-      Blake2b256(context) => (&mut **context).update(data),
-      Blake2b384(context) => (&mut **context).update(data),
+      Blake2b256(context) => Digest::update(&mut **context, data),
+      Blake2b384(context) => Digest::update(&mut **context, data),
       Blake2s(context) => Digest::update(&mut **context, data),
       Blake3(context) => Digest::update(&mut **context, data),
       Keccak224(context) => Digest::update(&mut **context, data),
       Keccak256(context) => Digest::update(&mut **context, data),
       Keccak384(context) => Digest::update(&mut **context, data),
       Keccak512(context) => Digest::update(&mut **context, data),
+      Md4(context) => Digest::update(&mut **context, data),
       Md5(context) => Digest::update(&mut **context, data),
       Ripemd160(context) => Digest::update(&mut **context, data),
       Sha1(context) => Digest::update(&mut **context, data),
@@ -232,6 +242,7 @@ impl Context {
       Sha256(context) => Digest::update(&mut **context, data),
       Sha384(context) => Digest::update(&mut **context, data),
       Sha512(context) => Digest::update(&mut **context, data),
+      Tiger(context) => Digest::update(&mut **context, data),
       Shake128(context) => (&mut **context).update(data),
       Shake256(context) => (&mut **context).update(data),
     };
@@ -253,14 +264,15 @@ impl Context {
 
     Ok(match self {
       Blake2b(context) => context.finalize(),
-      Blake2b256(context) => context.finalize_boxed(),
-      Blake2b384(context) => context.finalize_boxed(),
+      Blake2b256(context) => context.finalize(),
+      Blake2b384(context) => context.finalize(),
       Blake2s(context) => context.finalize(),
       Blake3(context) => context.finalize_boxed(length),
       Keccak224(context) => context.finalize(),
       Keccak256(context) => context.finalize(),
       Keccak384(context) => context.finalize(),
       Keccak512(context) => context.finalize(),
+      Md4(context) => context.finalize(),
       Md5(context) => context.finalize(),
       Ripemd160(context) => context.finalize(),
       Sha1(context) => context.finalize(),
@@ -272,6 +284,7 @@ impl Context {
       Sha256(context) => context.finalize(),
       Sha384(context) => context.finalize(),
       Sha512(context) => context.finalize(),
+      Tiger(context) => context.finalize(),
       Shake128(context) => context.finalize_boxed(length),
       Shake256(context) => context.finalize_boxed(length),
     })
@@ -292,14 +305,17 @@ impl Context {
     let length = length.unwrap_or_else(|| self.output_length());
     Ok(match self {
       Blake2b(context) => DynDigest::finalize_reset(context.as_mut()),
-      Blake2b256(context) => context.finalize_boxed_reset(),
-      Blake2b384(context) => context.finalize_boxed_reset(),
+      Blake2b256(context) => DynDigest::finalize_reset(context.as_mut()),
+      Blake2b384(context) => DynDigest::finalize_reset(context.as_mut()),
       Blake2s(context) => DynDigest::finalize_reset(context.as_mut()),
-      Blake3(context) => context.finalize_boxed_reset(length),
+      Blake3(context) => {
+        ExtendableOutputReset::finalize_boxed_reset(context.as_mut(), length)
+      }
       Keccak224(context) => DynDigest::finalize_reset(context.as_mut()),
       Keccak256(context) => DynDigest::finalize_reset(context.as_mut()),
       Keccak384(context) => DynDigest::finalize_reset(context.as_mut()),
       Keccak512(context) => DynDigest::finalize_reset(context.as_mut()),
+      Md4(context) => DynDigest::finalize_reset(context.as_mut()),
       Md5(context) => DynDigest::finalize_reset(context.as_mut()),
       Ripemd160(context) => DynDigest::finalize_reset(context.as_mut()),
       Sha1(context) => DynDigest::finalize_reset(context.as_mut()),
@@ -311,8 +327,13 @@ impl Context {
       Sha256(context) => DynDigest::finalize_reset(context.as_mut()),
       Sha384(context) => DynDigest::finalize_reset(context.as_mut()),
       Sha512(context) => DynDigest::finalize_reset(context.as_mut()),
-      Shake128(context) => context.finalize_boxed_reset(length),
-      Shake256(context) => context.finalize_boxed_reset(length),
+      Tiger(context) => DynDigest::finalize_reset(context.as_mut()),
+      Shake128(context) => {
+        ExtendableOutputReset::finalize_boxed_reset(context.as_mut(), length)
+      }
+      Shake256(context) => {
+        ExtendableOutputReset::finalize_boxed_reset(context.as_mut(), length)
+      }
     })
   }