+++ node/testdata/process_exit2.ts
@@ -1,0 +1,5 @@
+import "../global.ts";
+
+//deno-lint-ignore no-undef
+process.on("exit", () => console.log("exit"));
+process.exit();