+++ node/util.ts
@@ -1,6 +1,7 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
-import { promisify } from "./_util/_util_promisify.ts";
+import { promisify } from "./internal/util.js";
 import { callbackify } from "./_util/_util_callbackify.ts";
+import { debuglog } from "./_util/_debuglog.ts";
 import { deprecate } from "./internal/util.js";
 import {
   format,
@@ -13,12 +14,13 @@ import {
   ERR_OUT_OF_RANGE,
   errorMap
 } from "./internal/errors.ts";
-import * as types from "./internal/util/types.ts";
+import types from "./util/types.js";
 import { Buffer } from "./buffer.ts";
 import { isDeepStrictEqual } from "./internal/util/comparisons.ts";
 
 export {
   callbackify,
+  debuglog,
   deprecate,
   format,
   formatWithOptions,
@@ -218,11 +220,6 @@ function timestamp(): string {
 // deno-lint-ignore no-explicit-any
 function log(...args: any[]): void {
   console.log("%s - %s", timestamp(), format(...args));
-}
-
-// TODO(kt3k): implement debuglog correctly
-function debuglog() {
-  return console.log;
 }
 
 export default {