+++ node/util/types.js
@@ -1,0 +1,3 @@
+import * as types from "../internal/util/types.ts";
+
+export default { ...types };